Reference actions and epics typed
