+++ src/client/src/configureStore.ts
@@ -9,7 +9,7 @@ import { openfinEpic } from './openfinEpics'
 import { compositeStatusServiceEpic } from './operations/compositeStatusService'
 import { connectionStatusEpic } from './operations/connectionStatus'
 import { pricingServiceEpic } from './operations/pricing'
-import { referenceServiceEpic } from './referenceDataOperations'
+import { referenceServiceEpic } from './operations/referenceData'
 import { analyticsServiceEpic } from './ui/analytics'
 import { blotterEpic } from './ui/blotter/'
 import { popoutEpic } from './ui/common/popout/popoutEpic'
