+++ src/client/src/operations/currencyPairs/reducer.ts
@@ -1,5 +1,5 @@
 import { ACTION_TYPES as CONNECTION_ACTION_TYPES, DisconnectAction } from '../../operations/connectionStatus'
-import { ACTION_TYPES, ReferenceServiceAction } from '../../referenceDataOperations'
+import { ACTION_TYPES, ReferenceActions } from '../../operations/referenceData'
 import { CurrencyPair } from '../../types'
 import { toObject } from './utils'
 
@@ -11,7 +11,7 @@ const initialState: CurrencyPairState = {}
 
 export const currencyPairReducer = (
   state: CurrencyPairState = initialState,
-  action: ReferenceServiceAction | DisconnectAction
+  action: ReferenceActions | DisconnectAction
 ): CurrencyPairState => {
   switch (action.type) {
     case ACTION_TYPES.REFERENCE_SERVICE:
