+++ src/client/src/operations/referenceData/index.ts
@@ -1,0 +1,2 @@
+export { ACTION_TYPES, ReferenceActions } from './actions'
+export { referenceServiceEpic } from './epics'
